requirements.txt
bcrypt==3.2.0                        
Django==3.1.7               
django-cors-headers==3.7.0               
djangorestframework==3.12.2                       
Pillow==7.0.0          
psycopg2-binary==2.8.6                
python-decouple==3.4                 
pytz==2021.1